I solved Probelms From the 8th Course from 4 to 7

+++ Workspace/Workspace/C8 PS L4 P4 Number of Day-Hours-Minutes-Seconds In a Year.cpp
@@ -25,7 +25,7 @@ int NumberOfSecendsInYear(int Year)
 {
 	return NumberOfMinutsInYear(Year) * 60;
 }
-int ReadYear()
+short ReadYear()
 {
 	int Year;
 	cout << "\nPlease enter a year to check? ";
@@ -36,7 +36,7 @@ int ReadYear()
 
 int  main()
 {
-	int Year = ReadYear();
+	short Year = ReadYear();
 	
 
 		cout << "\nNumber of Days    in Year [" << Year << "] is " << NumberOfDaysInYear(Year);
